# openapi.yaml
...
openapi: 3.0.3
info:
    title: Subscription API
    version: 0.0.1
paths:
    /v1/subscription/auto-renew:
        post:
            tags:
                - Subscription
            description: 设置自动续费
            operationId: Subscription_SetAutoRenew
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/SetAutoRenewRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SetAutoRenewReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/auto-renew/process:
        post:
            tags:
                - Subscription
            description: 处理自动续费（用于定时任务）
            operationId: Subscription_ProcessAutoRenewals
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ProcessAutoRenewalsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProcessAutoRenewalsReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/cancel:
        post:
            tags:
                - Subscription
            description: 取消订阅
            operationId: Subscription_CancelSubscription
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CancelSubscriptionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CancelSubscriptionReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/expired/update:
        post:
            tags:
                - Subscription
            description: 批量更新过期订阅状态（用于定时任务）
            operationId: Subscription_UpdateExpiredSubscriptions
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateExpiredSubscriptionsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateExpiredSubscriptionsReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/expiring:
        get:
            tags:
                - Subscription
            description: 获取即将过期的订阅（用于定时任务）
            operationId: Subscription_GetExpiringSubscriptions
            parameters:
                - name: daysBeforeExpiry
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetExpiringSubscriptionsReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/history/{uid}:
        get:
            tags:
                - Subscription
            description: 获取订阅历史记录
            operationId: Subscription_GetSubscriptionHistory
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetSubscriptionHistoryReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/my/{uid}:
        get:
            tags:
                - Subscription
            description: 获取用户的订阅状态
            operationId: Subscription_GetMySubscription
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetMySubscriptionReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/order:
        post:
            tags:
                - Subscription
            description: 创建订阅订单 (调用 Payment Service)
            operationId: Subscription_CreateSubscriptionOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateSubscriptionOrderRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateSubscriptionOrderReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/pause:
        post:
            tags:
                - Subscription
            description: 暂停订阅
            operationId: Subscription_PauseSubscription
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PauseSubscriptionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PauseSubscriptionReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/payment/success:
        post:
            tags:
                - Subscription
            description: 支付回调处理 (通常由 Payment Service 或 MQ 调用)
            operationId: Subscription_HandlePaymentSuccess
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/HandlePaymentSuccessRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HandlePaymentSuccessReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/plans:
        get:
            tags:
                - Subscription
            description: 获取所有订阅套餐
            operationId: Subscription_ListPlans
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListPlansReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /v1/subscription/resume:
        post:
            tags:
                - Subscription
            description: 恢复订阅
            operationId: Subscription_ResumeSubscription
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ResumeSubscriptionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ResumeSubscriptionReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
...
        ProcessAutoRenewalsRequest:
            type: object
            properties:
                daysBeforeExpiry:
                    type: integer
                    format: int32
                dryRun:
                    type: boolean
            description: 自动续费处理
...
        SetAutoRenewReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
        SetAutoRenewRequest:
            type: object
            properties:
                uid:
                    type: string
                autoRenew:
                    type: boolean
            description: 自动续费设置
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
...
        UpdateExpiredSubscriptionsRequest:
            type: object
            properties: {}
            description: 批量更新过期订阅
...