# subscription_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: subscription.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationSubscriptionCancelSubscription = "/subscription.v1.Subscription/CancelSubscription"
const OperationSubscriptionCreateSubscriptionOrder = "/subscription.v1.Subscription/CreateSubscriptionOrder"
const OperationSubscriptionGetExpiringSubscriptions = "/subscription.v1.Subscription/GetExpiringSubscriptions"
const OperationSubscriptionGetMySubscription = "/subscription.v1.Subscription/GetMySubscription"
const OperationSubscriptionGetSubscriptionHistory = "/subscription.v1.Subscription/GetSubscriptionHistory"
const OperationSubscriptionHandlePaymentSuccess = "/subscription.v1.Subscription/HandlePaymentSuccess"
const OperationSubscriptionListPlans = "/subscription.v1.Subscription/ListPlans"
const OperationSubscriptionPauseSubscription = "/subscription.v1.Subscription/PauseSubscription"
const OperationSubscriptionProcessAutoRenewals = "/subscription.v1.Subscription/ProcessAutoRenewals"
const OperationSubscriptionResumeSubscription = "/subscription.v1.Subscription/ResumeSubscription"
const OperationSubscriptionSetAutoRenew = "/subscription.v1.Subscription/SetAutoRenew"
const OperationSubscriptionUpdateExpiredSubscriptions = "/subscription.v1.Subscription/UpdateExpiredSubscriptions"

type SubscriptionHTTPServer interface {
	// CancelSubscription 取消订阅
	CancelSubscription(context.Context, *CancelSubscriptionRequest) (*CancelSubscriptionReply, error)
	// CreateSubscriptionOrder 创建订阅订单 (调用 Payment Service)
	CreateSubscriptionOrder(context.Context, *CreateSubscriptionOrderRequest) (*CreateSubscriptionOrderReply, error)
	// GetExpiringSubscriptions 获取即将过期的订阅（用于定时任务）
	GetExpiringSubscriptions(context.Context, *GetExpiringSubscriptionsRequest) (*GetExpiringSubscriptionsReply, error)
	// GetMySubscription 获取用户的订阅状态
	GetMySubscription(context.Context, *GetMySubscriptionRequest) (*GetMySubscriptionReply, error)
	// GetSubscriptionHistory 获取订阅历史记录
	GetSubscriptionHistory(context.Context, *GetSubscriptionHistoryRequest) (*GetSubscriptionHistoryReply, error)
	// HandlePaymentSuccess 支付回调处理 (通常由 Payment Service 或 MQ 调用)
	HandlePaymentSuccess(context.Context, *HandlePaymentSuccessRequest) (*HandlePaymentSuccessReply, error)
	// ListPlans 获取所有订阅套餐
	ListPlans(context.Context, *ListPlansRequest) (*ListPlansReply, error)
	// PauseSubscription 暂停订阅
	PauseSubscription(context.Context, *PauseSubscriptionRequest) (*PauseSubscriptionReply, error)
	// ProcessAutoRenewals 处理自动续费（用于定时任务）
	ProcessAutoRenewals(context.Context, *ProcessAutoRenewalsRequest) (*ProcessAutoRenewalsReply, error)
	// ResumeSubscription 恢复订阅
	ResumeSubscription(context.Context, *ResumeSubscriptionRequest) (*ResumeSubscriptionReply, error)
	// SetAutoRenew 设置自动续费
	SetAutoRenew(context.Context, *SetAutoRenewRequest) (*SetAutoRenewReply, error)
	// UpdateExpiredSubscriptions 批量更新过期订阅状态（用于定时任务）
	UpdateExpiredSubscriptions(context.Context, *UpdateExpiredSubscriptionsRequest) (*UpdateExpiredSubscriptionsReply, error)
}

func RegisterSubscriptionHTTPServer(s *http.Server, srv SubscriptionHTTPServer) {
	r := s.Route("/")
	r.GET("/v1/subscription/plans", _Subscription_ListPlans0_HTTP_Handler(srv))
	r.GET("/v1/subscription/my/{uid}", _Subscription_GetMySubscription0_HTTP_Handler(srv))
	r.POST("/v1/subscription/order", _Subscription_CreateSubscriptionOrder0_HTTP_Handler(srv))
	r.POST("/v1/subscription/payment/success", _Subscription_HandlePaymentSuccess0_HTTP_Handler(srv))
	r.POST("/v1/subscription/cancel", _Subscription_CancelSubscription0_HTTP_Handler(srv))
	r.POST("/v1/subscription/pause", _Subscription_PauseSubscription0_HTTP_Handler(srv))
	r.POST("/v1/subscription/resume", _Subscription_ResumeSubscription0_HTTP_Handler(srv))
	r.GET("/v1/subscription/history/{uid}", _Subscription_GetSubscriptionHistory0_HTTP_Handler(srv))
	r.POST("/v1/subscription/auto-renew", _Subscription_SetAutoRenew0_HTTP_Handler(srv))
	r.GET("/v1/subscription/expiring", _Subscription_GetExpiringSubscriptions0_HTTP_Handler(srv))
	r.POST("/v1/subscription/expired/update", _Subscription_UpdateExpiredSubscriptions0_HTTP_Handler(srv))
	r.POST("/v1/subscription/auto-renew/process", _Subscription_ProcessAutoRenewals0_HTTP_Handler(srv))
}

func _Subscription_ListPlans0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListPlansRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionListPlans)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListPlans(ctx, req.(*ListPlansRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListPlansReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_GetMySubscription0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMySubscriptionRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionGetMySubscription)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMySubscription(ctx, req.(*GetMySubscriptionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMySubscriptionReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_CreateSubscriptionOrder0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateSubscriptionOrderRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionCreateSubscriptionOrder)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateSubscriptionOrder(ctx, req.(*CreateSubscriptionOrderRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateSubscriptionOrderReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_HandlePaymentSuccess0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in HandlePaymentSuccessRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionHandlePaymentSuccess)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.HandlePaymentSuccess(ctx, req.(*HandlePaymentSuccessRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*HandlePaymentSuccessReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_CancelSubscription0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CancelSubscriptionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionCancelSubscription)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CancelSubscription(ctx, req.(*CancelSubscriptionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CancelSubscriptionReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_PauseSubscription0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PauseSubscriptionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionPauseSubscription)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PauseSubscription(ctx, req.(*PauseSubscriptionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PauseSubscriptionReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_ResumeSubscription0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ResumeSubscriptionRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionResumeSubscription)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ResumeSubscription(ctx, req.(*ResumeSubscriptionRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ResumeSubscriptionReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_GetSubscriptionHistory0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetSubscriptionHistoryRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionGetSubscriptionHistory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetSubscriptionHistory(ctx, req.(*GetSubscriptionHistoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetSubscriptionHistoryReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_SetAutoRenew0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SetAutoRenewRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionSetAutoRenew)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SetAutoRenew(ctx, req.(*SetAutoRenewRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SetAutoRenewReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_GetExpiringSubscriptions0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetExpiringSubscriptionsRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionGetExpiringSubscriptions)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetExpiringSubscriptions(ctx, req.(*GetExpiringSubscriptionsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetExpiringSubscriptionsReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_UpdateExpiredSubscriptions0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateExpiredSubscriptionsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionUpdateExpiredSubscriptions)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateExpiredSubscriptions(ctx, req.(*UpdateExpiredSubscriptionsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateExpiredSubscriptionsReply)
		return ctx.Result(200, reply)
	}
}

func _Subscription_ProcessAutoRenewals0_HTTP_Handler(srv SubscriptionHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ProcessAutoRenewalsRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationSubscriptionProcessAutoRenewals)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ProcessAutoRenewals(ctx, req.(*ProcessAutoRenewalsRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ProcessAutoRenewalsReply)
		return ctx.Result(200, reply)
	}
}

type SubscriptionHTTPClient interface {
	// CancelSubscription 取消订阅
	CancelSubscription(ctx context.Context, req *CancelSubscriptionRequest, opts ...http.CallOption) (rsp *CancelSubscriptionReply, err error)
	// CreateSubscriptionOrder 创建订阅订单 (调用 Payment Service)
	CreateSubscriptionOrder(ctx context.Context, req *CreateSubscriptionOrderRequest, opts ...http.CallOption) (rsp *CreateSubscriptionOrderReply, err error)
	// GetExpiringSubscriptions 获取即将过期的订阅（用于定时任务）
	GetExpiringSubscriptions(ctx context.Context, req *GetExpiringSubscriptionsRequest, opts ...http.CallOption) (rsp *GetExpiringSubscriptionsReply, err error)
	// GetMySubscription 获取用户的订阅状态
	GetMySubscription(ctx context.Context, req *GetMySubscriptionRequest, opts ...http.CallOption) (rsp *GetMySubscriptionReply, err error)
	// GetSubscriptionHistory 获取订阅历史记录
	GetSubscriptionHistory(ctx context.Context, req *GetSubscriptionHistoryRequest, opts ...http.CallOption) (rsp *GetSubscriptionHistoryReply, err error)
	// HandlePaymentSuccess 支付回调处理 (通常由 Payment Service 或 MQ 调用)
	HandlePaymentSuccess(ctx context.Context, req *HandlePaymentSuccessRequest, opts ...http.CallOption) (rsp *HandlePaymentSuccessReply, err error)
	// ListPlans 获取所有订阅套餐
	ListPlans(ctx context.Context, req *ListPlansRequest, opts ...http.CallOption) (rsp *ListPlansReply, err error)
	// PauseSubscription 暂停订阅
	PauseSubscription(ctx context.Context, req *PauseSubscriptionRequest, opts ...http.CallOption) (rsp *PauseSubscriptionReply, err error)
	// ProcessAutoRenewals 处理自动续费（用于定时任务）
	ProcessAutoRenewals(ctx context.Context, req *ProcessAutoRenewalsRequest, opts ...http.CallOption) (rsp *ProcessAutoRenewalsReply, err error)
	// ResumeSubscription 恢复订阅
	ResumeSubscription(ctx context.Context, req *ResumeSubscriptionRequest, opts ...http.CallOption) (rsp *ResumeSubscriptionReply, err error)
	// SetAutoRenew 设置自动续费
	SetAutoRenew(ctx context.Context, req *SetAutoRenewRequest, opts ...http.CallOption) (rsp *SetAutoRenewReply, err error)
	// UpdateExpiredSubscriptions 批量更新过期订阅状态（用于定时任务）
	UpdateExpiredSubscriptions(ctx context.Context, req *UpdateExpiredSubscriptionsRequest, opts ...http.CallOption) (rsp *UpdateExpiredSubscriptionsReply, err error)
}

type SubscriptionHTTPClientImpl struct {
	cc *http.Client
}

func NewSubscriptionHTTPClient(client *http.Client) SubscriptionHTTPClient {
	return &SubscriptionHTTPClientImpl{client}
}

// CancelSubscription 取消订阅
func (c *SubscriptionHTTPClientImpl) CancelSubscription(ctx context.Context, in *CancelSubscriptionRequest, opts ...http.CallOption) (*CancelSubscriptionReply, error) {
	var out CancelSubscriptionReply
	pattern := "/v1/subscription/cancel"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSubscriptionCancelSubscription))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// CreateSubscriptionOrder 创建订阅订单 (调用 Payment Service)
func (c *SubscriptionHTTPClientImpl) CreateSubscriptionOrder(ctx context.Context, in *CreateSubscriptionOrderRequest, opts ...http.CallOption) (*CreateSubscriptionOrderReply, error) {
	var out CreateSubscriptionOrderReply
	pattern := "/v1/subscription/order"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSubscriptionCreateSubscriptionOrder))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetExpiringSubscriptions 获取即将过期的订阅（用于定时任务）
func (c *SubscriptionHTTPClientImpl) GetExpiringSubscriptions(ctx context.Context, in *GetExpiringSubscriptionsRequest, opts ...http.CallOption) (*GetExpiringSubscriptionsReply, error) {
	var out GetExpiringSubscriptionsReply
	pattern := "/v1/subscription/expiring"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSubscriptionGetExpiringSubscriptions))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetMySubscription 获取用户的订阅状态
func (c *SubscriptionHTTPClientImpl) GetMySubscription(ctx context.Context, in *GetMySubscriptionRequest, opts ...http.CallOption) (*GetMySubscriptionReply, error) {
	var out GetMySubscriptionReply
	pattern := "/v1/subscription/my/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSubscriptionGetMySubscription))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetSubscriptionHistory 获取订阅历史记录
func (c *SubscriptionHTTPClientImpl) GetSubscriptionHistory(ctx context.Context, in *GetSubscriptionHistoryRequest, opts ...http.CallOption) (*GetSubscriptionHistoryReply, error) {
	var out GetSubscriptionHistoryReply
	pattern := "/v1/subscription/history/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSubscriptionGetSubscriptionHistory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// HandlePaymentSuccess 支付回调处理 (通常由 Payment Service 或 MQ 调用)
func (c *SubscriptionHTTPClientImpl) HandlePaymentSuccess(ctx context.Context, in *HandlePaymentSuccessRequest, opts ...http.CallOption) (*HandlePaymentSuccessReply, error) {
	var out HandlePaymentSuccessReply
	pattern := "/v1/subscription/payment/success"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSubscriptionHandlePaymentSuccess))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListPlans 获取所有订阅套餐
func (c *SubscriptionHTTPClientImpl) ListPlans(ctx context.Context, in *ListPlansRequest, opts ...http.CallOption) (*ListPlansReply, error) {
	var out ListPlansReply
	pattern := "/v1/subscription/plans"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationSubscriptionListPlans))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// PauseSubscription 暂停订阅
func (c *SubscriptionHTTPClientImpl) PauseSubscription(ctx context.Context, in *PauseSubscriptionRequest, opts ...http.CallOption) (*PauseSubscriptionReply, error) {
	var out PauseSubscriptionReply
	pattern := "/v1/subscription/pause"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSubscriptionPauseSubscription))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ProcessAutoRenewals 处理自动续费（用于定时任务）
func (c *SubscriptionHTTPClientImpl) ProcessAutoRenewals(ctx context.Context, in *ProcessAutoRenewalsRequest, opts ...http.CallOption) (*ProcessAutoRenewalsReply, error) {
	var out ProcessAutoRenewalsReply
	pattern := "/v1/subscription/auto-renew/process"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSubscriptionProcessAutoRenewals))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ResumeSubscription 恢复订阅
func (c *SubscriptionHTTPClientImpl) ResumeSubscription(ctx context.Context, in *ResumeSubscriptionRequest, opts ...http.CallOption) (*ResumeSubscriptionReply, error) {
	var out ResumeSubscriptionReply
	pattern := "/v1/subscription/resume"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSubscriptionResumeSubscription))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SetAutoRenew 设置自动续费
func (c *SubscriptionHTTPClientImpl) SetAutoRenew(ctx context.Context, in *SetAutoRenewRequest, opts ...http.CallOption) (*SetAutoRenewReply, error) {
	var out SetAutoRenewReply
	pattern := "/v1/subscription/auto-renew"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSubscriptionSetAutoRenew))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateExpiredSubscriptions 批量更新过期订阅状态（用于定时任务）
func (c *SubscriptionHTTPClientImpl) UpdateExpiredSubscriptions(ctx context.Context, in *UpdateExpiredSubscriptionsRequest, opts ...http.CallOption) (*UpdateExpiredSubscriptionsReply, error) {
	var out UpdateExpiredSubscriptionsReply
	pattern := "/v1/subscription/expired/update"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationSubscriptionUpdateExpiredSubscriptions))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}